requirements.txt
langchain
langchain-community
langchain-pinecone
langchain-openai
openai
llama-index
pinecone-client
google-api-python-client
fastapi
uvicorn[standard]
python-dotenv